Refactor user containers
This removes the subscribing container around every route, and adds a
subscribing container on top of the AppLayout, which passes all the data
to all its children all the time. It's then up to each route to pick
the data that it's interest in. the first component that does this is
the RequestContainer, which wraps all routes which are related to a
single loanRequest.