EP-77 Insert loans, borrowers and properties with proper userId
If userId is set, you're inserting for someone else's account, so
you're an admin. If userId is undefined, you're inserting for
yourself, so the userId will be determined by meteor. If userId is
null, it should remain that way, as the user will add it to his
account later on.